+++ playx/songfinder.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """It is an abstract module for searching songs"""
 
-from youtube import search_youtube
+from .youtube import search_youtube
 
 def search(song):
     """Search the song in youtube."""
