+++ requirements.txt
@@ -1,3 +1,3 @@
-youtube_dl==2018.3.20
-requests==2.18.4
+youtube_dl==2018.6.25
+requests==2.19.1
 beautifulsoup4==4.6.0
